+++ CHANGES.rst
@@ -8,6 +8,7 @@ Development
 - Improve documentation
 - Improve voltage divider settings for reading the battery level
 - Package the release bundle with the same directory layout as the sandbox
+- Add foundation for having button events through ESP32 touch pads
 
 
 2019-06-17 0.4.0